fix: address PR comments
