+++ modules/migration/src/main/java/com/sap/xsk/migration/api/XSKMigrationRestService.java
@@ -26,9 +26,9 @@ import com.sap.xsk.migration.api.dto.DeliveryUnitData;
 import com.sap.xsk.migration.api.dto.ExecuteMigrationRequestBody;
 import com.sap.xsk.migration.api.dto.MigrationRequestBody;
 import com.sap.xsk.migration.api.dto.MigrationResponseBody;
-import com.sap.xsk.migration.neo.db.DeliveryUnitsExportConfig;
-import com.sap.xsk.migration.neo.db.DeliveryUnitsExporter;
-import com.sap.xsk.migration.neo.db.DeliveryUnitsProvider;
+import com.sap.xsk.migration.neo.db.hana.DeliveryUnitsExportConfig;
+import com.sap.xsk.migration.neo.db.hana.DeliveryUnitsExporter;
+import com.sap.xsk.migration.neo.db.hana.DeliveryUnitsProvider;
 import com.sap.xsk.migration.neo.db.hana.connectivity.HanaConnector;
 import com.sap.xsk.migration.neo.db.hana.connectivity.HanaConnectorArgs;
 import com.sap.xsk.migration.neo.db.hana.connectivity.HanaConnectorRes;
@@ -39,7 +39,6 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.Authorization;
-import org.apache.commons.lang3.SystemUtils;
 import org.eclipse.dirigible.commons.api.service.AbstractRestService;
 import org.eclipse.dirigible.commons.api.service.IRestService;
 import org.eclipse.dirigible.core.workspace.service.WorkspacesCoreService;
