+++ modules/migration/src/main/java/com/sap/xsk/migration/neo/db/hana/ConnectionProvider.java
@@ -4,7 +4,7 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
-class HanaConnectionProvider {
+public class ConnectionProvider {
 
   public Connection getConnection(String url, String user, String password) throws SQLException {
     return DriverManager.getConnection(url, user, password);
