+++ modules/migration/src/main/java/com/sap/xsk/migration/neo/sdk/command/databases/ListDatabasesSdkCommand.java
@@ -13,22 +13,25 @@ package com.sap.xsk.migration.neo.sdk.command.databases;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-import com.sap.xsk.migration.neo.sdk.command.AbstractSdkCommand;
+import com.sap.xsk.migration.neo.sdk.command.SdkCommand;
 import com.sap.xsk.migration.neo.sdk.command.SdkCommandGenericArgs;
 import com.sap.xsk.migration.neo.sdk.command.SdkCommandParsedOutput;
 import com.sap.xsk.migration.tooling.MigrationToolExecutor;
+
 import javax.inject.Inject;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-class ListDatabasesSdkCommand extends AbstractSdkCommand<SdkCommandGenericArgs, ListDatabasesSdkCommandRes> {
+public class ListDatabasesSdkCommand implements SdkCommand<SdkCommandGenericArgs, ListDatabasesSdkCommandRes> {
 
   private static final String LIST_DATABASES_COMMAND_NAME = "list-dbs";
 
+  private final MigrationToolExecutor migrationToolExecutor;
+
   @Inject
   public ListDatabasesSdkCommand(MigrationToolExecutor migrationToolExecutor) {
-    super(migrationToolExecutor);
+    this.migrationToolExecutor = migrationToolExecutor;
   }
 
   @Override
