+++ modules/migration/src/main/java/com/sap/xsk/migration/neo/sdk/command/tunnel/CloseDatabaseTunnelSdkCommand.java
@@ -11,29 +11,27 @@
  */
 package com.sap.xsk.migration.neo.sdk.command.tunnel;
 
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-import com.sap.xsk.migration.neo.sdk.command.AbstractSdkCommand;
-import com.sap.xsk.migration.neo.sdk.command.SdkCommandParsedOutput;
+import com.sap.xsk.migration.neo.sdk.command.SdkCommand;
 import com.sap.xsk.migration.tooling.MigrationToolExecutor;
+
 import javax.inject.Inject;
 import java.util.List;
 
-class CloseDatabaseTunnelSdkCommand extends AbstractSdkCommand<CloseDatabaseTunnelSdkCommandArgs, CloseDatabaseTunnelSdkCommandRes> {
+public class CloseDatabaseTunnelSdkCommand implements SdkCommand<CloseDatabaseTunnelSdkCommandArgs, Void> {
 
   private static final String CLOSE_DATABASE_TUNNEL_COMMAND_NAME = "close-db-tunnel";
 
+  private final MigrationToolExecutor migrationToolExecutor;
+
   @Inject
   public CloseDatabaseTunnelSdkCommand(MigrationToolExecutor migrationToolExecutor) {
-    super(migrationToolExecutor);
+    this.migrationToolExecutor = migrationToolExecutor;
   }
 
   @Override
-  public CloseDatabaseTunnelSdkCommandRes execute(CloseDatabaseTunnelSdkCommandArgs commandArgs) {
+  public Void execute(CloseDatabaseTunnelSdkCommandArgs commandArgs) {
     List<String> commandAndArgs = createProcessCommandAndArguments(commandArgs, CLOSE_DATABASE_TUNNEL_COMMAND_NAME);
-    String rawCommandOutput = migrationToolExecutor.executeMigrationTool(NEO_SDK_DIRECTORY, commandAndArgs);
-    SdkCommandParsedOutput<CloseDatabaseTunnelSdkCommandRes> parsedCommandOutput = new Gson().fromJson(rawCommandOutput, new TypeToken<SdkCommandParsedOutput<CloseDatabaseTunnelSdkCommandRes>>() {
-    }.getType());
-    return parsedCommandOutput.getResult();
+    migrationToolExecutor.executeMigrationTool(NEO_SDK_DIRECTORY, commandAndArgs);
+    return null;
   }
 }
