--- modules/migration/src/main/java/com/sap/xsk/migration/tooling/SystemEnvironment.java
@@ -1,8 +1,0 @@
-package com.sap.xsk.migration.tooling;
-
-class SystemEnvironment {
-
-  public String getEnvironmentVariableValue(String environmentVariableName) {
-    return System.getenv(environmentVariableName);
-  }
-}
