+++ modules/migration/src/test/java/com/sap/xsk/migration/api/XSKMigrationRestServiceTest.java
@@ -5,9 +5,9 @@ import com.sap.xsk.migration.api.dto.HanaData;
 import com.sap.xsk.migration.api.dto.MigrationRequestBody;
 import com.sap.xsk.migration.api.dto.MigrationResponseBody;
 import com.sap.xsk.migration.api.dto.NeoData;
-import com.sap.xsk.migration.neo.db.DeliveryUnit;
-import com.sap.xsk.migration.neo.db.DeliveryUnitsExporter;
-import com.sap.xsk.migration.neo.db.DeliveryUnitsProvider;
+import com.sap.xsk.migration.neo.db.hana.DeliveryUnit;
+import com.sap.xsk.migration.neo.db.hana.DeliveryUnitsExporter;
+import com.sap.xsk.migration.neo.db.hana.DeliveryUnitsProvider;
 import com.sap.xsk.migration.neo.db.hana.connectivity.HanaConnector;
 import com.sap.xsk.migration.neo.db.hana.connectivity.HanaConnectorRes;
 import com.sap.xsk.migration.neo.sdk.command.SdkCommand;
